app: add ui_meta keys for controlling some ui properties dynamically
Add support for the "sensitive", "visible", "label", and "description"
property keys for gegl ops.

The "sensitive" and "visible" keys should be boolean GUM expressions,
and the "label" and "description" keys should be string GUM
expressions, controlling the sensitivity, visibility, label, and tool-
tip of the corresponding prop gui widget, respectively.